Basic clean main menu form
